requirements.txt
# Core dependencies
Flask
Flask-PyMongo
pymongo
Flask-CORS
pandas
python-dateutil
fuzzywuzzy
python-Levenshtein
python-dotenv

# Code Quality
pylint
vulture

# Testing dependencies
pytest
pytest-flask
playwright
pytest-playwright
pytest-cov
